indexer/pyproject.toml
[project]
name = "indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.18",
    "memory-profiler>=0.61.0",
    "numpy>=2.2.6",
    "openai>=1.82.0",
    "prometheus-client>=0.22.1",
    "psutil>=7.0.0",
    "pydantic-settings>=2.9.1",
    "pytest-benchmark>=5.1.0",
    "python-jose>=3.4.0",
    "qdrant-client>=1.14.2",
    "redis[asyncio]>=6.2.0",
]
